# Remove naked pointers from the VM