@maximedenes 2018-04-20
@ppedrot What is the status of this? Did you test the gains?