@ppedrot 2018-04-20
@maximedenes I didn't manage to make the pendulum OPAM installation work yet, I get mysterious errors when the packages work on my machine. Might be due to some old OPAM bug, I have no idea. But I'll retry when I have more time.

Note also that this patch doesn't hurt the current implementation and enforces an invariant that is desirable in OCaml code in general. I don't know how much they're pushing for it, but I think that even if nobody uses this compiler flag it still makes sense to preserve it in Coq if there is no disadvantage. (Also, this might simplify the proper registering of VM code in the GC.)