@ppedrot 2018-04-23
@maximedenes I reran the compcert bench twice, and got respectively 
```
┌──────────────┬───────────────────────┬─────────────────────────────────────┬─────────────────────────────────────┬─────────────────────────┬─────────────────┐
│              │     user time [s]     │             CPU cycles              │          CPU instructions           │  max resident mem [KB]  │   mem faults    │
│              │                       │                                     │                                     │                         │                 │
│ package_name │    NEW    OLD PDIFF   │           NEW           OLD PDIFF   │           NEW           OLD PDIFF   │     NEW     OLD PDIFF   │ NEW OLD PDIFF   │
├──────────────┼───────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┼─────────────────────────┼─────────────────┤
│ coq-compcert │ 875.80 867.81 +0.92 % │ 2427366966541 2401614242347 +1.07 % │ 3482406399942 3473830266057 +0.25 % │ 1319144 1328152 -0.68 % │   0   0  +nan % │
└──────────────┴───────────────────────┴─────────────────────────────────────┴─────────────────────────────────────┴─────────────────────────┴─────────────────┘
│ coq-compcert │ 875.51 867.32 +0.94 % │ 2425210964851 2400801118549 +1.02 % │ 3481605374489 3474332731664 +0.21 % │ 1319016 1328184 -0.69 % │   0   0  +nan % │
└──────────────┴───────────────────────┴─────────────────────────────────────┴─────────────────────────────────────┴─────────────────────────┴─────────────────┘
```

So some minor slowdown seems to be going on.