@ppedrot 2018-04-23
@maximedenes Oops, sorry, I must be tired, and answered in the wrong PR. This is indeed for #7085.