@maximedenes 2018-04-23
> That's noise level, you can't conclude anything aside from no major
> changes from it.

I'm not so sure. Usually, CPU cycles don't vary like that with just noise. But not impossible, obviously.