@maximedenes 2018-04-30
I'm not sure @bgregoir and I have the knowledge required to review this particular patch. Maybe @silene can help? If not, I'll look more into it, but it will take some time.