@ppedrot 2018-04-30
I'll rebase this patch for an easier review.