@silene 2018-05-09
Overall it seems fine. That being said, I am curious as to why you didn't use the same trick as `accumulate`  for `makeaccu` and `pushpop` (i.e. a pseudo header). It seems like it would have avoided a bunch of indirections and made the code simpler.