@ppedrot 2018-05-09
@silene This could have been possible for these specific functions, but you do have to have indirections in general around `tcode`. The reason is that code can be arbitrarily shifted by `offset_closure` and the like, and when creating a closure a code pointer can actually point in an arbitrary place in the bytecode, where the contiguous memory zone need not be a block header.