@ppedrot 2018-05-09
@maximedenes BTW, native_compute is also incompatible with no-naked-pointers. The trick it uses to encode the accumulator block breaks this invariant because it is a block of tag 0 that contains a code pointer. I am actually pretty much surprised this is not supposed to be UB in OCaml.