@silene 2018-05-09
> This could have been possible for these specific functions

Yes, I was specifically talking about these functions.