@ppedrot 2018-05-09
@silene Actually, I misremember: it's the other way around. The accumulate block can appear without the wrapper which is why it needs a header, but the two other don't. So we're actually saving memory in this case, making this scheme uniform would require adding a dummy header. And we always need the abstract tag because bytecode can point in general to arbitrary memory, as mentioned above.

So this PR is somewhat optimal.