@silene 2018-05-10
> So this PR is somewhat optimal.

I don't quite follow. You have two possibilities to fix the naked pointer issue (and you implemented both). Either you add a dummy header to the static block, or you allocate an abstract block to wrap around the pointer to the static block. The dummy header costs one memory word, while the abstract block costs two memory words. So the dummy header is the cheapest one, but it can only be used when the pointer points at the start of the static block, which is the case here.

Anyway, the optimality does not matter much here, since there will never be more than a handful of these static blocks.